Move registration module in to sign_up.ex
